pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "orin-system-setup"
version = "0.1.0"
description = "System setup tools for NVIDIA Orin devices"
readme = "README.md"
requires-python = ">=3.9"

dependencies = [
    "websockets>=15.0.1",
    "boto3>=1.35.0",
    "requests>=2.31.0",
    "pyyaml>=6.0.1",
]

[tool.setuptools.packages.find]
include = ["OTA*", "WIFI*"]

[dependency-groups]
dev = [
    "ruff>=0.9.3",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.5.0",
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pre-commit>=3.0.0",
    "pyright>=1.1.406",
]

[tool.black]
line-length = 88
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["OTA", "WIFI"]
known_third_party = ["flask", "requests", "boto3", "websockets"]

[tool.ruff]
line-length = 88
indent-width = 4
src = ["OTA", "WIFI"]
exclude = ["__pycache__", "*.pyc", "build", "dist", ".git"]

[tool.ruff.lint]
select = ["E", "F"]
ignore = ["E203", "E501"]
fixable = ["ALL"]
unfixable = []
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"
